Write value of argument "--config" to attribute "config_file"
Without this change the value of the argument "--config" given on the command
line is ignored completely, but a new default config file gets gcreated. With
this change the config file specified on the command line is used correctly.

As far as I can see in the code the same seems to apply to the command line
argument "--hooks" and the attribute "hooks_file".